myproject/requirement.txt
aerich==0.7.2
aiosqlite==0.17.0
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
certifi==2023.11.17
click==8.1.7
colorama==0.4.6
dictdiffer==0.9.0
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
idna==3.6
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.10
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.5.2
pydantic-extra-types==2.2.0
pydantic-settings==2.1.0
pydantic_core==2.14.5
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomlkit==0.12.3
tortoise-orm==0.20.0
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
